+++ http/src/main/scala/com/twitter/finatra/http/filters/TraceIdMDCFilter.scala
@@ -2,6 +2,7 @@ package com.twitter.finatra.http.filters
 
 import com.twitter.finagle.tracing.Trace
 import com.twitter.finagle.{Service, SimpleFilter}
+import com.twitter.inject.logging.MDCKeys
 import com.twitter.util.Future
 import javax.inject.Singleton
 import org.slf4j.MDC
@@ -17,11 +18,12 @@ import org.slf4j.MDC
 class TraceIdMDCFilter[Req, Rep] extends SimpleFilter[Req, Rep] {
 
   override def apply(request: Req, service: Service[Req, Rep]): Future[Rep] = {
-    MDC.put("traceId", Trace.id.traceId.toString)
+    val trace = Trace()
+    MDC.put(MDCKeys.TraceId, trace.id.traceId.toString)
     // If sampling decision is not made yet
     // consider the trace not sampled for this span scope
-    MDC.put("traceSampled", Trace.id._sampled.getOrElse(false).toString)
-    MDC.put("traceSpanId", Trace.id.spanId.toString)
+    MDC.put(MDCKeys.TraceSampled, trace.id._sampled.getOrElse(false).toString)
+    MDC.put(MDCKeys.TraceSpanId, trace.id.spanId.toString)
     service(request)
   }
 }