+++ thrift/src/main/scala/com/twitter/finatra/thrift/filters/TraceIdMDCFilter.scala
@@ -2,6 +2,7 @@ package com.twitter.finatra.thrift.filters
 
 import com.twitter.finagle.{Filter, Service}
 import com.twitter.finagle.tracing.Trace
+import com.twitter.inject.logging.MDCKeys
 import com.twitter.util.Future
 import javax.inject.Singleton
 import org.slf4j.MDC
@@ -20,11 +21,12 @@ class TraceIdMDCFilter extends Filter.TypeAgnostic {
       request: T,
       service: Service[T, U]
     ): Future[U] = {
-      MDC.put("traceId", Trace.id.traceId.toString)
+      val trace = Trace()
+      MDC.put(MDCKeys.TraceId, trace.id.traceId.toString)
       // If sampling decision is not made yet
       // consider the trace not sampled for this span scope
-      MDC.put("traceSampled", Trace.id._sampled.getOrElse(false).toString)
-      MDC.put("traceSpanId", Trace.id.spanId.toString)
+      MDC.put(MDCKeys.TraceSampled, trace.id._sampled.getOrElse(false).toString)
+      MDC.put(MDCKeys.TraceSpanId, trace.id.spanId.toString)
       service(request)
     }
   }